+++ instrumentation/ratpack-1.4/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/ratpack/ServerRegistryInstrumentation.java
@@ -12,9 +12,12 @@ import static net.bytebuddy.matcher.ElementMatchers.named;
 
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import java.util.Map;
+import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
+import ratpack.handling.HandlerDecorator;
+import ratpack.registry.Registry;
 
 public class ServerRegistryInstrumentation implements TypeInstrumentation {
 
@@ -27,6 +30,15 @@ public class ServerRegistryInstrumentation implements TypeInstrumentation {
   public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
     return singletonMap(
         isMethod().and(isStatic()).and(named("buildBaseRegistry")),
-        ServerRegistryAdvice.class.getName());
+        ServerRegistryInstrumentation.class.getName() + "$BuildAdvice");
+  }
+
+  public static class BuildAdvice {
+    @Advice.OnMethodExit(suppress = Throwable.class)
+    public static void injectTracing(@Advice.Return(readOnly = false) Registry registry) {
+      registry =
+          registry.join(
+              Registry.builder().add(HandlerDecorator.prepend(TracingHandler.INSTANCE)).build());
+    }
   }
 }
